Update tsp java client depencencies

+++ app/src/main/java/org/eclipse/trace/coordinator/core/xy/XyService.java
@@ -32,7 +32,7 @@ public class XyService {
 			final String outputId,
 			final Body<GetXyModelRequestDto> body) {
 
-		final List<Integer> traceServerRequestedItems = List.copyOf(body.getParameters().getRequestedItems()).stream()
+		final List<Long> traceServerRequestedItems = List.copyOf(body.getParameters().getRequestedItems()).stream()
 				.filter(traceServer::isValidEncodeEntryId)
 				.map(traceServer::decodeEntryId)
 				.collect(Collectors.toList());
